move dynamic price logic to hook
